+++ src/FreeDSx/Ldap/Server/Backend/Storage/Adapter/InMemoryStorage.php
@@ -76,4 +76,9 @@ final class InMemoryStorage implements EntryStorageInterface
     {
         $operation($this);
     }
+
+    public function namingContexts(): array
+    {
+        return $this->namingContextsFromArray($this->entries);
+    }
 }